refactor: #399 피드백 반영-1
- FakeController, GithubResponses 패키지 위치 변경
테스트 영역에서 관리되므로 test 패키지에 위치시켰다.

- GithubProperties 필드에 private 접근제한자 사용
값을 읽을 수 있도록 getter 메서드 추가
