+++ src/main/java/nextstep/member/application/config/GithubProperties.java
@@ -6,31 +6,55 @@ import org.springframework.boot.context.properties.ConstructorBinding;
 @ConstructorBinding
 @ConfigurationProperties(prefix = "github")
 public class GithubProperties {
-    public final Client client;
-    public final Url url;
+    private final Client client;
+    private final Url url;
 
     public GithubProperties(Client client, Url url) {
         this.client = client;
         this.url = url;
     }
 
+    public Client getClient() {
+        return client;
+    }
+
+    public Url getUrl() {
+        return url;
+    }
+
     public static class Client {
-        public final String id;
-        public final String secret;
+        private final String id;
+        private final String secret;
 
         public Client(String id, String secret) {
             this.id = id;
             this.secret = secret;
         }
+
+        public String getId() {
+            return id;
+        }
+
+        public String getSecret() {
+            return secret;
+        }
     }
 
     public static class Url {
-        public final String accessToken;
-        public final String profile;
+        private final String accessToken;
+        private final String profile;
 
         public Url(String accessToken, String profile) {
             this.accessToken = accessToken;
             this.profile = profile;
+        }
+
+        public String getAccessToken() {
+            return accessToken;
+        }
+
+        public String getProfile() {
+            return profile;
         }
     }
 }
